Only print client states we see, not all of them.
There are more client states enumerated than there really are
states in the handshake. This was done to accomodate concurrent
versions of TLS in the enum, but it makes it hard to read.
The table is more compact now since we only print the states
that were seen in the handshake.